Fixed some map bug, inputforms does run again.
